Optimize processing batches of VNC packets
If there are multiple VNC packets in the recv buffer when we go to handle
them, we can avoid performing a memmove after every event and instead just keep
track of an event offset. All reads for a given event are done relative to that
offset instead of at an absolute index into the recv buffer.

Often the entire recv buffer is a contiguous group of events, so no copying is
necessary at all. Even when copying is necessary we do it only once per recv.

Also clean up some comments and the computation of the zlib frame size. We don't
need to do a full division since we can count the frames by just shifting off
the bottom 16 bits of the data size.